Renamed service-location-service-channel-styore
