+++ gutenberg/blocks/ts/src/index.ts
@@ -1,3 +1,3 @@
-import './store';
+import './service-location-service-channel-store';
 import './service-block';
 import './service-location-service-channel-block';
